action-library: Extract .getToken() into a new environment module
Change-type: minor
