+++ lib/action-library/index.js
@@ -20,6 +20,7 @@ const skhema = require('skhema')
 const syncContext = require('./sync-context')
 const sync = require('../sync')
 const logger = require('../logger')
+const environment = require('../environment')
 
 const slugify = (string) => {
 	return string
@@ -34,7 +35,7 @@ module.exports = {
 			timestamp: request.timestamp,
 			context: syncContext.fromWorkerContext(
 				context, logger.context.systemContext, session),
-			token: syncContext.getToken(card.data.source)
+			token: environment.getIntegrationToken(card.data.source)
 		})
 	},
 	'action-integration-github-mirror-event': async (session, context, card, request) => {
@@ -43,7 +44,7 @@ module.exports = {
 			integration: 'github',
 			context: syncContext.fromWorkerContext(
 				context, logger.context.systemContext, session),
-			token: syncContext.getToken('github')
+			token: environment.getIntegrationToken('github')
 		})
 	},
 	'action-integration-front-mirror-event': async (session, context, card, request) => {
@@ -52,7 +53,7 @@ module.exports = {
 			integration: 'front',
 			context: syncContext.fromWorkerContext(
 				context, logger.context.systemContext, session),
-			token: syncContext.getToken('front')
+			token: environment.getIntegrationToken('front')
 		})
 	},
 	'action-create-card': async (session, context, card, request) => {
