+++ lib/action-library/sync-context.js
@@ -16,35 +16,6 @@
 
 const logger = require('../logger').getLogger(__filename)
 
-exports.getToken = (integration) => {
-	if (integration === 'github') {
-		const result = {
-			api: process.env.INTEGRATION_GITHUB_TOKEN || null,
-			signature: process.env.INTEGRATION_GITHUB_SIGNATURE_KEY || null
-		}
-
-		if (!result.api || !result.signature) {
-			return null
-		}
-
-		return result
-	}
-
-	if (integration === 'front') {
-		const result = {
-			api: process.env.INTEGRATION_FRONT_TOKEN
-		}
-
-		if (!result.api) {
-			return null
-		}
-
-		return result
-	}
-
-	return null
-}
-
 exports.fromWorkerContext = (workerContext, context, session) => {
 	return {
 		log: {
