+++ lib/core/assert.js
@@ -16,12 +16,13 @@
 
 const _ = require('lodash')
 const assert = require('assert')
+const environment = require('../environment')
 
 // This will be evaluated once, and then
 // cached by the module system, so future
 // requires won't have to do this check
 
-if (process.env.NODE_ENV === 'production') {
+if (environment.isProduction()) {
 	for (const fn of Object.keys(assert)) {
 		exports[fn] = _.noop
 	}
