+++ lib/logger/index.js
@@ -2,6 +2,7 @@ const _ = require('lodash')
 const path = require('path')
 const winston = require('winston')
 const loggerContext = require('./context')
+const environment = require('../environment')
 const BASE_PATH = path.join(__dirname, '..', '..')
 
 const config = {
@@ -76,9 +77,9 @@ module.exports.getLogger = _.memoize((filename) => {
 		.slice(1)
 		.join(path.sep)
 
-	const loglevel = process.env.CI || process.env.AVA_PATH
+	const loglevel = environment.isTest()
 		? 'warn'
-		: (config.logLevel[name] || process.env.LOGLEVEL || 'debug')
+		: (config.logLevel[name] || environment.getLogLevel())
 	const myFormat = winston.format.printf((info) => {
 		const additionalArgs = info.additionalArgs
 		const id = _.get(info, [ 'ctx', 'id' ], 'MISSING-TRANSACTION')
