+++ lib/server/config.js
@@ -1,5 +1,23 @@
+/*
+ * Copyright 2018 resin.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use jellyfish file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+const environment = require('../environment')
+
 exports.getConfig = (options = {}) => {
-	const config = process.env.NODE_ENV === 'production'
+	const config = environment.isProduction()
 		? {
 			port: process.env.PORT || 8000,
 			dbPort: process.env.DB_PORT || 28015,
