+++ lib/server/error-reporter.js
@@ -16,10 +16,11 @@
 
 const Raven = require('raven')
 const logger = require('../logger').getLogger(__filename)
+const environment = require('../environment')
 
 class Reporter {
 	constructor () {
-		if (process.env.NODE_ENV === 'production') {
+		if (environment.isProduction()) {
 			if (process.env.SENTRY_DSN_SERVER) {
 				Raven.config(process.env.SENTRY_DSN_SERVER).install()
 				this.initialized = true
