+++ lib/server/routes.js
@@ -18,8 +18,8 @@ const Bluebird = require('bluebird')
 const randomstring = require('randomstring')
 const sync = require('../sync')
 const queue = require('../queue')
+const environment = require('../environment')
 const logger = require('../logger').getLogger(__filename)
-const syncContext = require('../action-library/sync-context')
 const fs = require('fs')
 const _ = require('lodash')
 const multer = require('multer')
@@ -178,7 +178,8 @@ exports.bindRoutes = (jellyfish, worker, app, socketServer) => {
 			source: request.params.provider
 		})
 
-		const integrationToken = syncContext.getToken(request.params.provider)
+		const integrationToken = environment.getIntegrationToken(
+			request.params.provider)
 
 		if (!sync.isValidExternalEventRequest(
 			integrationToken, request.params.provider, request.rawBody, request.headers)) {
