+++ lib/server/web-server.js
@@ -18,6 +18,7 @@ const bodyParser = require('body-parser')
 const compression = require('compression')
 const portfinder = require('portfinder')
 const logger = require('../logger').getLogger(__filename)
+const environment = require('../environment')
 const express = require('express')
 const http = require('http')
 const socketIo = require('socket.io')
@@ -94,7 +95,7 @@ exports.createWebServer = (config) => {
 		server.on('error', async (error) => {
 			// If the server is running in a test environment and the
 			// port is in use, try binding to a higher port number.
-			if (error.code === 'EADDRINUSE' && process.env.NODE_ENV === 'test') {
+			if (error.code === 'EADDRINUSE' && !environment.isProduction()) {
 				// The original listener has to be removed, otherwise it will be
 				// triggered when the server finds an available port
 				logger.warn(null, 'Port already in use', {
